+++ packages/vm/src/instance.rs
@@ -390,7 +390,7 @@ mod tests {
         let instance = mock_instance(&CONTRACT, &[]);
 
         instance
-            .call_function0("interface_version_6", &[])
+            .call_function0("interface_version_7", &[])
             .expect("error calling function");
     }
 
@@ -533,7 +533,7 @@ mod tests {
 
                 (type (func))
                 (func (type 0) nop)
-                (export "interface_version_6" (func 0))
+                (export "interface_version_7" (func 0))
                 (export "instantiate" (func 0))
                 (export "allocate" (func 0))
                 (export "deallocate" (func 0))
@@ -551,7 +551,7 @@ mod tests {
 
                 (type (func))
                 (func (type 0) nop)
-                (export "interface_version_6" (func 0))
+                (export "interface_version_7" (func 0))
                 (export "instantiate" (func 0))
                 (export "allocate" (func 0))
                 (export "deallocate" (func 0))